[mongo 수정] 1. Id 추가, 변경

+++ src/main/java/com/movie_api/db/repo/UserRepo.java
@@ -4,8 +4,9 @@ import com.movie_api.db.collection.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 
+import java.math.BigInteger;
 import java.util.List;
-public interface UserRepo extends MongoRepository<User, Long> {
+public interface UserRepo extends MongoRepository<User, BigInteger> {
     // 비밀번호는 출력하지 않음
     @Query(value = "{}", fields = "{'pw':0}")
     List<User> findAll();
